Created separate Card render and punchButton
